requirements.txt
streamlit
huggingface-hub
sentence-transformers
scikit-learn
numpy
python-dotenv
pytest
pre-commit
black
ruff
langchain
langchain-community
langchain-huggingface
langchain-core
langgraph
chromadb
ipython
pypdf
langsmith